+++ package.json
@@ -4,6 +4,7 @@
   "description": "Provides a mock RPC API given an open rpc schema",
   "main": "server.js",
   "dependencies": {
+    "@open-rpc/schema-utils-js": "^1.0.1",
     "jayson": "^2.1.1",
     "json-schema-faker": "^0.5.0-rc16",
     "json-schema-ref-parser": "^6.0.3",
